Made the enemies move better.
Still needs some work. Enemies still move somewhat randomly and quickly. I also need to work on their SURROUNDING behavior a little bit
