+++ Capstone/Assets/Scripts/StateMachine/EnemyBehaviors/Actions/EnemyATTACK.cs
@@ -80,7 +80,7 @@ namespace Bladesmiths.Capstone
             _sword.GetComponent<BoxCollider>().enabled = true;
             timer = 0f;
             timerMax = 1f;
-            _enemy.attackTimerMax = Random.Range(0.75f, 3f);
+            _enemy.attackTimerMax = Random.Range(0.75f, 2f);
             _enemy.canMove = true;
         }
