+++ Capstone/Assets/Scripts/StateMachine/EnemyBehaviors/Actions/EnemyATTACK_SHIELD.cs
@@ -87,7 +87,7 @@ namespace Bladesmiths.Capstone
             _sword.GetComponent<BoxCollider>().enabled = true;
             timer = 0f;
             timerMax = 1f;
-            _enemy.attackTimerMax = Random.Range(0.75f, 3f);
+            _enemy.attackTimerMax = Random.Range(0.75f, 2f);
         }
 
         public override void OnEnd()
