+++ Capstone/Assets/Scripts/StateMachine/EnemyBehaviors/Conditionals/WithinDistance.cs
@@ -9,7 +9,6 @@ namespace Bladesmiths.Capstone
     public class WithinDistance : Conditional
     {
         private Transform player;
-        public float fieldOfView;
         public float distance;
         public float distVec;
 
@@ -21,33 +20,14 @@ namespace Bladesmiths.Capstone
 
         public override TaskStatus OnUpdate()
         {
-            float newFOV = fieldOfView;
-            if (GetComponent<Enemy>().InCombat)
-            {
-                newFOV = 360;
-            }
-
             distVec = Vector3.Distance(player.position, transform.position);
 
-            if (InSight(player, newFOV) && (distVec <= distance))
+            if (distVec <= distance)
             {
                 distVec = 10000;
                 return TaskStatus.Success;
             }
             return TaskStatus.Failure;
         }
-
-        public override void OnEnd()
-        {
-            base.OnEnd();
-        }
-
-        public bool InSight(Transform target, float fov)
-        {
-            Vector3 dir = target.position - transform.position;
-            return Vector3.Angle(dir, transform.forward) < fov;
-        }
-
-
     }
 }
